refactor: 패키지 정리
- domain을 기준으로 패키지를 다시 정리하였음.
- domain과 global(여러 도메인에서 사용되는 클래스)로 큰 구분
- domain에는 각각의 도메인(여기서는 book)이 패키지로 존재
- book 내에는 web(controller, web dto, interceptor, handler), service(application service, infra service, service dto), domain, repository, exception이 존재
- global에는 common(공통 클래스), utils(정적 도구 클래스), error(공통 exception 객체 및 exception handler), logger(log에 대한 interceptor), config(설정 파일)이 존재
- 테스트도 이에 맞게 패키지 변경
