--- src/main/java/com/booksdiary/config/JpaAuditingConfig.java
@@ -1,10 +1,0 @@
-package com.booksdiary.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaAuditingConfig {
-
-}
