--- src/main/java/com/booksdiary/config/WebMvcConfig.java
@@ -1,23 +1,0 @@
-package com.booksdiary.config;
-
-import com.booksdiary.controller.common.LogInterceptor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-    private final LogInterceptor logInterceptor;
-
-    public WebMvcConfig(LogInterceptor logInterceptor) {
-        this.logInterceptor = logInterceptor;
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(logInterceptor).addPathPatterns("/**")
-                .excludePathPatterns("/css/**")
-                .excludePathPatterns("/js/**")
-                .excludePathPatterns("/index.html");
-    }
-}
