--- src/main/java/com/booksdiary/controller/common/ErrorCode.java
@@ -1,29 +1,0 @@
-package com.booksdiary.controller.common;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-
-@Getter
-@JsonFormat(shape = JsonFormat.Shape.OBJECT)
-public enum ErrorCode {
-
-    // Common
-    INVALID_INPUT_VALUE(400, "C001", " Invalid Input Value"),
-    METHOD_NOT_ALLOWED(405, "C002", " Invalid Input Value"),
-    ENTITY_NOT_FOUND(400, "C003", " Entity Not Found"),
-    INTERNAL_SERVER_ERROR(500, "C004", "Server Error"),
-    INVALID_TYPE_VALUE(400, "C005", " Invalid Type Value"),
-    HANDLE_ACCESS_DENIED(403, "C006", "Access is Denied");
-
-    // TODO: 2021/02/22 여기에서 로그인, 도메인, 기타 등등의 예외 코드를 추가할 수 있다.
-
-    private final String code;
-    private final String message;
-    private final int status;
-
-    ErrorCode(final int status, final String code, final String message) {
-        this.status = status;
-        this.message = message;
-        this.code = code;
-    }
-}
