--- src/main/java/com/booksdiary/controller/common/ErrorResponse.java
@@ -1,83 +1,0 @@
-package com.booksdiary.controller.common;
-
-import lombok.Getter;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Getter
-public class ErrorResponse {
-    private final String message;
-    private final int status;
-    private final List<FieldError> errors;
-    private final String code;
-
-
-    private ErrorResponse(final ErrorCode code, final List<FieldError> errors) {
-        this.message = code.getMessage();
-        this.status = code.getStatus();
-        this.errors = errors;
-        this.code = code.getCode();
-    }
-
-    private ErrorResponse(final ErrorCode code) {
-        this.message = code.getMessage();
-        this.status = code.getStatus();
-        this.code = code.getCode();
-        this.errors = new ArrayList<>();
-    }
-
-
-    public static ErrorResponse of(final ErrorCode code, final BindingResult bindingResult) {
-        return new ErrorResponse(code, FieldError.of(bindingResult));
-    }
-
-    public static ErrorResponse of(final ErrorCode code) {
-        return new ErrorResponse(code);
-    }
-
-    public static ErrorResponse of(final ErrorCode code, final List<FieldError> errors) {
-        return new ErrorResponse(code, errors);
-    }
-
-    public static ErrorResponse of(MethodArgumentTypeMismatchException e) {
-        final String value = e.getValue() == null ? "" : e.getValue().toString();
-        final List<ErrorResponse.FieldError> errors = ErrorResponse.FieldError.of(e.getName(), value, e.getErrorCode());
-        return new ErrorResponse(ErrorCode.INVALID_TYPE_VALUE, errors);
-    }
-
-
-    @Getter
-    public static class FieldError {
-        private final String field;
-        private final String value;
-        private final String reason;
-
-        private FieldError(final String field, final String value, final String reason) {
-            this.field = field;
-            this.value = value;
-            this.reason = reason;
-        }
-
-        public static List<FieldError> of(final String field, final String value, final String reason) {
-            List<FieldError> fieldErrors = new ArrayList<>();
-            fieldErrors.add(new FieldError(field, value, reason));
-            return fieldErrors;
-        }
-
-        private static List<FieldError> of(final BindingResult bindingResult) {
-            final List<org.springframework.validation.FieldError> fieldErrors = bindingResult.getFieldErrors();
-            return fieldErrors.stream()
-                    .map(error -> new FieldError(
-                            error.getField(),
-                            error.getRejectedValue() == null ? "" : error.getRejectedValue().toString(),
-                            error.getDefaultMessage()))
-                    .collect(Collectors.toList());
-        }
-    }
-
-
-}
