--- src/main/java/com/booksdiary/domain/BookRepository.java
@@ -1,6 +1,0 @@
-package com.booksdiary.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookRepository extends JpaRepository<Book, Long> {
-}
