+++ src/main/java/com/booksdiary/domain/book/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.booksdiary.domain.book.repository;
+
+import com.booksdiary.domain.book.domain.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
