--- src/main/java/com/booksdiary/exception/BusinessException.java
@@ -1,19 +1,0 @@
-package com.booksdiary.exception;
-
-import com.booksdiary.controller.common.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class BusinessException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public BusinessException(String message, ErrorCode errorCode) {
-        super(message);
-        this.errorCode = errorCode;
-    }
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
