--- src/main/java/com/booksdiary/exception/EntityNotFoundException.java
@@ -1,10 +1,0 @@
-package com.booksdiary.exception;
-
-import com.booksdiary.controller.common.ErrorCode;
-
-public class EntityNotFoundException extends BusinessException {
-
-    public EntityNotFoundException(String message) {
-        super(message, ErrorCode.ENTITY_NOT_FOUND);
-    }
-}
