--- src/main/java/com/booksdiary/exception/InvalidValueException.java
@@ -1,14 +1,0 @@
-package com.booksdiary.exception;
-
-import com.booksdiary.controller.common.ErrorCode;
-
-public class InvalidValueException extends BusinessException {
-
-    public InvalidValueException(String value) {
-        super(value, ErrorCode.INVALID_INPUT_VALUE);
-    }
-
-    public InvalidValueException(String value, ErrorCode errorCode) {
-        super(value, errorCode);
-    }
-}
