+++ src/main/java/com/booksdiary/global/config/JpaAuditingConfig.java
@@ -1,0 +1,10 @@
+package com.booksdiary.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaAuditingConfig {
+
+}
