+++ src/main/java/com/booksdiary/global/error/exception/BusinessException.java
@@ -1,0 +1,19 @@
+package com.booksdiary.global.error.exception;
+
+import com.booksdiary.global.error.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public BusinessException(String message, ErrorCode errorCode) {
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
