+++ src/main/java/com/booksdiary/global/error/exception/EntityNotFoundException.java
@@ -1,0 +1,10 @@
+package com.booksdiary.global.error.exception;
+
+import com.booksdiary.global.error.ErrorCode;
+
+public class EntityNotFoundException extends BusinessException {
+
+    public EntityNotFoundException(String message) {
+        super(message, ErrorCode.ENTITY_NOT_FOUND);
+    }
+}
