+++ src/main/java/com/booksdiary/global/error/exception/InvalidValueException.java
@@ -1,0 +1,14 @@
+package com.booksdiary.global.error.exception;
+
+import com.booksdiary.global.error.ErrorCode;
+
+public class InvalidValueException extends BusinessException {
+
+    public InvalidValueException(String value) {
+        super(value, ErrorCode.INVALID_INPUT_VALUE);
+    }
+
+    public InvalidValueException(String value, ErrorCode errorCode) {
+        super(value, errorCode);
+    }
+}
